pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=42", "wheel"]

[project]
authors = [
  {name = "Vadim Kozyrevskiy", email = "vadikko2@mail.ru"},
  {name = "Dmitry Kutlubaev", email = "kutlubaev00@mail.ru"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "pydantic==2.*",
  "orjson==3.9.15",
  "di[anyio]==0.79.2",
  "sqlalchemy[asyncio]==2.0.*",
  "retry-async==0.1.4",
  "python-dotenv==1.0.1",
  "dependency-injector>=4.48.2"
]
description = "Python CQRS pattern implementation"
maintainers = [{name = "Vadim Kozyrevskiy", email = "vadikko2@mail.ru"}]
name = "python-cqrs"
readme = "README.md"
requires-python = ">=3.10"
version = "4.4.1"

[project.optional-dependencies]
dev = [
  # Develope tools
  "pycln==2.5.0",
  "pre-commit==3.8.0",
  "pyright==1.1.377",
  "ruff==0.6.2",
  # Tests
  "aiokafka==0.10.0",
  "pytest~=7.4.2",
  "pytest-asyncio~=0.21.1",
  "pytest-env==0.6.2",
  "cryptography==42.0.2",
  "asyncmy==0.2.9",
  "requests>=2.32.5"
]
examples = [
  "fastapi==0.109.*",
  "uvicorn==0.32.0",
  "faststream[kafka]==0.5.28",
  "faker>=37.12.0"
]
kafka = [
  "aiokafka==0.10.0",
  # for SchemaRegistry
  "confluent-kafka==2.6.0"
]
protobuf = ["protobuf==4.25.5"]
rabbit = [
  "aio-pika==9.3.0"
]

[project.urls]
Documentation = "https://vadikko2.github.io/python-cqrs-mkdocs/"
Issues = "https://github.com/vadikko2/python-cqrs/issues"
Repository = "https://github.com/vadikko2/python-cqrs"

[tool.pytest.ini_options]
addopts = "--junit-xml=report.xml"
asyncio_mode = "auto"
junit_family = "xunit1"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
